fix railtie dependency to insert MiniProfiler after ActionDispatch::Static
It seems that MiniProfiler stopped running on start after the pull
request that adds MiniProfiler to the middleware based on where
other Middleware was in the stack. It seems MiniProfiler needs to be
added during initialization not after or else it won't run. From the
research I did I think it would be better to add MiniProfiler before
ActionDispatch::Static instead of checking against the current stack.
That way if Rack::Deflater is alse added before ActionDispatch::Static
then Rack::Delflater will still be before MiniProfiler.