Changes RedisStore's collection of unviewed ids from an unsorted set to a sorted set
This was done in order to:

1. Improve performance when clearing expired page struct ids from the set of unviewed ids
2. Ensure that the result of `RedisStore#get_unviewed_ids` maintains a reverse-chronological sort order

For 100,000 records, the performance impact of `RedisStore#get_unviewed_ids` has been reduced from ~0.5s per request to ~0.05s.

Unsorted set methods called within `RedisStore` have been updated to use sorted set methods. When a record is set as "unviewed", its time-to-live is looked up and added to the current time and the sum is used for scoring the elements within the set of unviewed ids. When `RedisStore#get_unviewed_ids` is called, all records in the unviewed ids set that have a score less than `Time.now.to_i` are removed from the set. This is intended to keep the set of unviewed ids in sync with the records those ids represent.

In order to reliably look up the TTL for each result, `RedisStore` now expects that the record has been saved prior to being set as unviewed. This also prevents invalid ids from being inserted into the set of unviewed ids. `MiniProfiler::Profiler#call` was adjusted so that it now calls `@storage.save` prior to calling `@storage.set_unviewed`. This does not appear to have any negative consequences.

Tests have been updated in accordance with these changes. Of note, a time-dependent test has been introduced in order to accurately simulate an expiring record scenario in redis. This adds ~4s to the test suite.