Ensures that `RedisStore#get_unviewed_ids` returns only ids for existing records (#292)
* Ensures that `RedisStore#get_unviewed_ids` returns only ids for existing records

When using rack-mini-profiler configured with `RedisStore`, I began encountering 404 responses when `includes.js` made POST requests to `/mini-profiler-resources/results` after 24 hours of use. These 404 errors were due to the store being unable to load records matching the ids sent by the client because they had expired. The client receives these ids via the `X-MiniProfiler-Ids` header, which is provided by `RedisStore#get_unviewed_ids`.

`RedisStore#get_unviewed_ids` had the possibility of containing references to redis records that were expired. This is because the list of unviewed ids and the individual `PageStruct` records could become out of sync; individual records have their expiration set once when initially saved, but the list of unviewed ids has its expiration renewed every time a record is added to the list. The result is that the records would expire, but the list would continue to reference those records.

To ensure that `RedisStore#get_unviewed_ids` only returns ids with matching existing records, it now iterates through the list of unviewed ids and removes members that do not exist.

A test has been added to `redis_store_spec.rb` to expose this issue. Additional cleanup work has been performed in this spec; notably, the store db is now flushed prior to each spec. This revealed some broken tests that I've attempted to clean up and fix.

* Moves `RedisStore`'s "unviewed ids" synchronization logic to lua

Per @SamSaffron's suggestion, removing expired ids within `RedisStore` is now done as part of a Lua command. This cuts down on round-tripping data between redis and ruby and should be more performant.

* Changes RedisStore's collection of unviewed ids from an unsorted set to a sorted set

This was done in order to:

1. Improve performance when clearing expired page struct ids from the set of unviewed ids
2. Ensure that the result of `RedisStore#get_unviewed_ids` maintains a reverse-chronological sort order

For 100,000 records, the performance impact of `RedisStore#get_unviewed_ids` has been reduced from ~0.5s per request to ~0.05s.

Unsorted set methods called within `RedisStore` have been updated to use sorted set methods. When a record is set as "unviewed", its time-to-live is looked up and added to the current time and the sum is used for scoring the elements within the set of unviewed ids. When `RedisStore#get_unviewed_ids` is called, all records in the unviewed ids set that have a score less than `Time.now.to_i` are removed from the set. This is intended to keep the set of unviewed ids in sync with the records those ids represent.

In order to reliably look up the TTL for each result, `RedisStore` now expects that the record has been saved prior to being set as unviewed. This also prevents invalid ids from being inserted into the set of unviewed ids. `MiniProfiler::Profiler#call` was adjusted so that it now calls `@storage.save` prior to calling `@storage.set_unviewed`. This does not appear to have any negative consequences.

Tests have been updated in accordance with these changes. Of note, a time-dependent test has been introduced in order to accurately simulate an expiring record scenario in redis. This adds ~4s to the test suite.