Allow filtering of stack frames in flamegraphs
For some larger Rails applications, the flamegraph can be pretty
overwhelming and full of non-application-related stack frames. While
that information is often valuable, at other times it's really helpful
to be able to see a simpler view of a smaller subset of the entire
process.

This leverages the changes added to flamegraph in
https://github.com/SamSaffron/flamegraph/pull/21 and adds a
rack-mini-profiler configuration to pass through to flamegraph to
achieve the above goal.