Patch for undefined MiniProfiler.templates
With the new approach to compiled templates, rack-mini-profiler loads
`includes.js` to start the system on every page load (full http request or
fetch). As part of its initialization, it loads the compiled templates in
`vendor.js`. When performing a turbolinks navigation, it will trigger both
a POST request to collect results and a full restart of the system. This
can result in `MiniProfiler.templates` being undefined in the result collection
request.