FEATURE: Use speedscope to display flamegraphs and switch from the flamegraph gem to stackprof (#475)
Mini Profiler now ships with an embedded copy of speedscope which
provides a very nice experience for viewing flamegraphs.

Flamegraph has been replaced with the stackprof gem. It's an optional
dependency, so if you use the flamegraphs feature, you'll need to add
the `stackprof` gem to your `Gemfile` and require it. If you don't use
the feature, you don't need to add anything and Mini Profiler will work
just fine.