Ensure compatibility with Rack 2.x and 3.x (#653)
Some parts of the code are changing the headers from the Rack request,
but are using capitalized headers when doing so. If the underlying
object is not a bare hash but either a `Rack::Utils::HeaderHash` or a
`Rack::Headers` object, then everything will work as expected. But
sometimes the headers object can be a bare hash (that can be the case
when Rails is serving static files) and if Rack 3 is in use, things can
break.

This patch ensures the headers names are compliant with both versions of
Rack. (mainly by using things like `Rack::CONTENT_TYPE` instead of
writing directly `Content-Type`/`content-type`)